+++ 5/1_9_courbes_orbitales.rs
@@ -82,7 +82,7 @@ fn calculate_points(model: &mut Model) {
     let n = model.settings.curve.n as f32;
     let t1 = model.settings.curve.t1 as f32;
     let t2 = model.settings.curve.t2 as f32;
-    let r1 = model.settings.curve.r1 as f32;
+    let r1 = model.settings.curve.r1;
     let k1 = model.settings.curve.k1 as f32;
     let k2 = model.settings.curve.k2 as f32;
 
@@ -93,7 +93,7 @@ fn calculate_points(model: &mut Model) {
         let a2 = 2.0 * PI * i / n * t2;
 
         let x = r1 * (k1 * a1).cos() + r2 * a2.cos();
-        let y = r1 * (k2 as f32 * a1).sin() + r2 * a2.sin();
+        let y = r1 * (k2 * a1).sin() + r2 * a2.sin();
 
         model.points.push(pt2(x, y));
     }
